Heal Start now checks if all units are alive already
 Changes to be committed:
	modified:   ALSpringBackend/src/main/java/com/ardaslegends/data/domain/Army.java
	modified:   ALSpringBackend/src/main/java/com/ardaslegends/data/service/ArmyService.java
	modified:   ALSpringBackend/src/main/java/com/ardaslegends/data/service/dto/army/UpdateArmyDto.java
	modified:   ALSpringBackend/src/main/java/com/ardaslegends/data/service/exceptions/army/ArmyServiceException.java
	modified:   ALSpringBackend/src/test/java/com/ardaslegends/presentation/api/ArmyRestControllerTest.java
	modified:   ALSpringBackend/src/test/java/com/ardaslegends/repository/MovementRepositoryTest.java
	modified:   ALSpringBackend/src/test/java/com/ardaslegends/service/ArmyServiceTest.java
	modified:   ALSpringBackend/src/test/java/com/ardaslegends/service/MovementServiceTest.java
	modified:   ALSpringBackend/src/test/java/com/ardaslegends/service/PathfinderTest.java

+++ ALSpringBackend/src/main/java/com/ardaslegends/data/service/dto/army/UpdateArmyDto.java
@@ -1,4 +1,4 @@
 package com.ardaslegends.data.service.dto.army;
 
-public record UpdateArmyDto(String executorDiscordId, String armyName, Integer freeTokens) {
+public record UpdateArmyDto(String executorDiscordId, String armyName, Double freeTokens) {
 }
